+++ R/sandwichMethods.R
@@ -50,7 +50,7 @@ bread.singleRStaticCountData <- function(x,...) {
 #' @author Piotr Chlebicki, Maciej Beręsewicz
 #' 
 #' @description S3 method for \code{vcovHC} to handle \code{singleRStaticCountData} class objects. 
-#' Works exactly like \code{vcov.default} the only difference being that this method handles vector generalised linear models.
+#' Works exactly like \code{vcovHC.default} the only difference being that this method handles vector generalised linear models.
 #' Updating the covariance matrix in variance/standard error estimation for population size estimator can be done via [singleRcapture::redoPopEstimation()]
 #'
 #' @param x a fitted \code{singleRStaticCountData} class object.